Make energy value a Double

+++ src/main/scala/com/agecomp/energy/Components.scala
@@ -3,5 +3,5 @@ package com.agecomp.energy
 import com.agecomp.Component
 import com.agecomp.InputComponent
 
-class EnergyComponent(var value: Int) extends Component {
+class EnergyComponent(var value: Double) extends Component {
 }
